requirements.txt
Django==3.2.3
mysqlclient==2.0.3
django-admin-searchable-dropdown==1.2
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
django-environ==0.9.0